requirements.txt
amqp==5.2.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
dj-database-url==2.2.0
django-environ==0.11.2
django-on-heroku==1.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
eventlet==0.36.1
fastjsonschema==2.20.0
gunicorn==22.0.0
ipython==8.12.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyterlab-pygments==0.3.0
kombu==5.3.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
pandocfilters==1.5.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
pymongo==4.7.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sentry-sdk==2.5.1
soupsieve==2.5
tinycss2==1.3.0
urllib3==2.2.1
vine==5.1.0
webencodings==0.5.1
whitenoise==6.7.0
yarg==0.1.9